+++ src/mongo/transport/service_executor_synchronous.h
@@ -68,7 +68,7 @@ private:
     stdx::condition_variable _shutdownCondition;
 
     AtomicWord<size_t> _numRunningWorkerThreads{0};
-    unsigned _numHardwareCores{0};
+    size_t _numHardwareCores{0};
 };
 
 }  // namespace transport
